# specs/002-roles-permissions-system/openapi.yaml
openapi: 3.0.3
info:
  title: WasteFlow Roles & Permissions API
  description: |
    Comprehensive roles and permissions management system for WasteFlow application.

    This API supports:
    - Multi-tenant role-based access control (RBAC)
    - Attribute-based access control (ABAC) policies
    - Consultant multi-tenant context switching
    - Temporary permission grants with auto-expiration
    - Comprehensive audit logging for regulatory compliance
    - Custom role creation for enterprise clients
    - Permission request workflows

    ## Authentication
    All endpoints require JWT Bearer authentication obtained via SPID/CIE login.
    The JWT contains user identity, tenant context, and SPID fiscal code for audit traceability.

    ## Multi-Tenancy
    - Schema-per-tenant architecture with Row-Level Security (RLS)
    - Tenant context extracted from JWT claims
    - Consultants can switch tenant context via dedicated endpoints
    - All operations scoped to current tenant unless explicitly cross-tenant (consultants only)

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Permission check endpoints: 1000 requests/minute per user (high-frequency caching)
    - Audit log queries: 20 requests/minute per user (expensive operations)
    - Bulk operations: 10 requests/minute per user

    ## Compliance
    - All audit logs retained for 10+ years per D.Lgs. 152/2006
    - Immutable audit trail with cryptographic integrity
...

  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /roles:
    get:
      tags: [Roles]
      summary: List all roles for current tenant
      description: |
        Returns all available roles including system-defined and custom roles.
        System roles (Admin, Operator, Viewer, etc.) cannot be modified.
        Custom roles are tenant-specific and can be edited/deleted if not assigned.
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [system, custom, all]
            default: all
          description: Filter by role type
        - name: search
          in: query
          schema:
            type: string
          description: Search role name or description (case-insensitive)
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRolesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 100/minute

    post:
      tags: [Custom Roles]
      summary: Create custom role
      description: |
        Creates a new custom role with granular permission matrix.
        Only available to Super Admin users.
        Enterprise tier feature - requires appropriate subscription.
      operationId: createCustomRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Custom role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - requires Super Admin role or Enterprise tier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: FORBIDDEN
                message: Custom role creation requires Super Admin role and Enterprise tier subscription
                statusCode: 403
                details:
                  requiredRole: SUPER_ADMIN
                  requiredTier: ENTERPRISE
                  currentTier: STANDARD
        '409':
          description: Conflict - role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 10/minute
  /roles/{roleId}:
    get:
      tags: [Roles]
      summary: Get role details
      description: |
        Returns complete role information including:
        - Permission matrix
        - Number of assigned users
        - Creation/modification history
        - Preview of capabilities
      operationId: getRoleById
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      responses:
        '200':
          description: Role details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 100/minute

    patch:
      tags: [Custom Roles]
      summary: Update custom role
      description: |
        Updates custom role permissions or metadata.
        System roles cannot be modified.
        Permission changes affect all users with this role immediately.
        Audit log entry created with before/after snapshots.
      operationId: updateCustomRole
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - cannot modify system roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute

    delete:
      tags: [Custom Roles]
      summary: Delete custom role
      description: |
        Deletes custom role if no users are currently assigned.
        System roles cannot be deleted.
        If users are assigned, returns error with reassignment guidance.
      operationId: deleteCustomRole
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      responses:
        '204':
          description: Role deleted successfully
        '400':
          description: Cannot delete - users still assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: ROLE_IN_USE
                message: This role is assigned to 12 users. Reassign them first before deleting.
                statusCode: 400
                details:
                  assignedUserCount: 12
                  assignedUserIds: ["user-1", "user-2", "..."]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute
  /roles/{roleId}/permissions:
    get:
      tags: [Roles]
      summary: Get role permissions matrix
      description: |
        Returns detailed permission matrix for role organized by module.
        Useful for permission preview before role assignment.
      operationId: getRolePermissions
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [module, action, scope]
            default: module
          description: How to group permissions in response
      responses:
        '200':
          description: Permission matrix retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionMatrixResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 100/minute
...
  /users/{userId}/roles/{roleId}:
    delete:
      tags: [Roles]
      summary: Revoke role from user
      description: |
        Removes role assignment from user.
        Prevents removal if this would leave tenant with no administrators.
        Creates audit log entry with revocation reason.
        Invalidates user's permission cache immediately.
      operationId: revokeRoleFromUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/RoleIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  minLength: 10
                  maxLength: 500
                  description: Reason for revocation (mandatory for audit)
              required: [reason]
      responses:
        '204':
          description: Role revoked successfully
        '400':
          description: Cannot revoke - would leave tenant without admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: LAST_ADMIN_PROTECTION
                message: At least one administrator is required per organization
                statusCode: 400
                details:
                  currentAdminCount: 1
                  requiredMinimum: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via SPID/CIE authentication.
        Include in Authorization header: `Bearer <token>`

        JWT contains claims:
        - userId: User unique identifier
        - tenantId: Current tenant context
        - fiscalCode: SPID fiscal code for audit
        - roles: User's role IDs
        - permissions: Cached permission set
        - exp: Token expiration timestamp
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
        default: createdAt
      description: Field to sort by
    SortOrderParam:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
    RoleIdParam:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Role unique identifier
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User unique identifier
...
  schemas:
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Role unique identifier
        name:
          type: string
          description: Role name (unique per tenant)
          example: Site Manager Milano
        description:
          type: string
          description: Human-readable role description
          example: Manages waste operations for Milano facility
        isSystemRole:
          type: boolean
          description: System role (cannot be modified/deleted)
        permissions:
          type: array
          items:
            type: string
          description: Array of permission strings (format resource:action:scope)
          example: ["fir:create:facility", "fir:read:all", "registry:write:facility"]
        assignedUserCount:
          type: integer
          description: Number of users currently with this role
        facilityScoped:
          type: boolean
          description: Whether role can be facility-scoped
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
          description: User who created this custom role
      required: [id, name, description, isSystemRole, permissions, assignedUserCount, createdAt]
    RoleDetailResponse:
      allOf:
        - $ref: '#/components/schemas/RoleResponse'
        - type: object
          properties:
            permissionsByModule:
              type: object
              description: Permissions organized by module for clear display
              additionalProperties:
                type: object
                properties:
                  module:
                    type: string
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/PermissionDetail'
              example:
                fir:
                  module: Formulari Rifiuti (FIR)
                  permissions:
                    - resource: fir
                      action: create
                      scope: facility
                      description: Create new FIR for assigned facilities
                registry:
                  module: Registry Management
                  permissions:
                    - resource: registry
                      action: read
                      scope: all
                      description: View all registry entries
            assignedUsers:
              type: array
              items:
                $ref: '#/components/schemas/UserRoleSummary'
              description: Summary of users with this role (first 10)
            changeHistory:
              type: array
              items:
                $ref: '#/components/schemas/RoleChangeRecord'
              description: Recent modification history (last 10 changes)
    PermissionDetail:
      type: object
      properties:
        resource:
          type: string
          description: Resource type
          example: fir
        action:
          type: string
          description: Action on resource
          enum: [create, read, update, delete, approve, sign, export]
        scope:
          type: string
          description: Permission scope
          enum: [own, facility, all]
        description:
          type: string
          description: Human-readable permission description (Italian)
        sensitive:
          type: boolean
          description: High-risk permission requiring extra scrutiny
      required: [resource, action, scope, description]
    UserRoleSummary:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        facilityIds:
          type: array
          items:
            type: string
            format: uuid
          description: Facility scope (if applicable)
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
          nullable: true
    RoleChangeRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        changeType:
          type: string
          enum: [CREATE, UPDATE, DELETE, ASSIGN, REVOKE]
        changedBy:
          type: string
          format: uuid
        changedByName:
          type: string
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
        beforeSnapshot:
          type: object
          description: State before change (JSON)
        afterSnapshot:
          type: object
          description: State after change (JSON)
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Unique role name within tenant
          example: Site Manager Milano
        description:
          type: string
          minLength: 10
          maxLength: 500
          description: Clear description of role purpose
          example: Manages waste operations for Milano facility with full FIR and registry access
        permissions:
          type: array
          items:
            type: string
            pattern: '^[a-z_]+:(create|read|update|delete|approve|sign|export):(own|facility|all)$'
          minItems: 1
          maxItems: 100
          description: Array of permission strings
          example: ["fir:create:facility", "fir:read:all", "registry:write:facility"]
        facilityScoped:
          type: boolean
          default: false
          description: Whether this role can be facility-scoped
      required: [name, description, permissions]
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          minLength: 10
          maxLength: 500
        permissions:
          type: array
          items:
            type: string
        reason:
          type: string
          minLength: 10
          maxLength: 500
          description: Reason for modification (mandatory for audit)
      required: [reason]
    PaginatedRolesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]

    # ===========================
    # USER ROLE ASSIGNMENT SCHEMAS
    # ===========================
...
    PermissionMatrixResponse:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        permissionsByModule:
          type: object
          additionalProperties:
            type: object
            properties:
              module:
                type: string
              permissions:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionDetail'
        totalPermissions:
          type: integer
        sensitivePermissions:
          type: array
          items:
            type: string
          description: High-risk permissions in this role
      required: [roleId, roleName, permissionsByModule, totalPermissions]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: PERMISSION_DENIED
        message:
          type: string
          description: Human-readable error message (Italian)
          example: Non hai il permesso necessario per eseguire questa azione
        statusCode:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint that generated error
      required: [error, message, statusCode, timestamp]
    ValidationErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message
        value:
          description: Invalid value provided
        constraint:
          type: string
          description: Validation constraint violated
      required: [field, message]
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Authentication required. Please login with SPID/CIE.
            statusCode: 401
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
    ForbiddenError:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Non hai il permesso necessario per eseguire questa azione
            statusCode: 403
            details:
              requiredPermission: roles:create
              currentRole: OPERATOR
              contactAdministrator: admin@example.com
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Role non trovato
            statusCode: 404
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles/123e4567-e89b-12d3-a456-426614174000
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  validationErrors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed
            statusCode: 400
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
            validationErrors:
              - field: name
                message: Role name must be at least 3 characters
                value: ab
                constraint: minLength
              - field: permissions
                message: At least one permission is required
                constraint: minItems
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.
            statusCode: 429
            details:
              limit: 100
              window: 60 seconds
              retryAfter: 45
            timestamp: 2025-10-31T14:30:00.000Z
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds until retry is allowed
          schema:
            type: integer

# ===========================
# WEBHOOKS
# ===========================
...